🔨  Refactor to smaller components
